move common types and errors out of scaffold
