+++ internal/service/contentprovider/securityconfig.go
@@ -3,8 +3,8 @@ package contentprovider
 import (
 	"fmt"
 
-	"github.com/kyma-project/modulectl/internal/scaffold/common/errors"
-	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
+	"github.com/kyma-project/modulectl/internal/common/errors"
+	"github.com/kyma-project/modulectl/internal/common/types"
 )
 
 type SecurityConfig struct {
