+++ internal/service/filegenerator/filegenerator.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 
-	commonerrors "github.com/kyma-project/modulectl/internal/scaffold/common/errors"
-	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
+	commonerrors "github.com/kyma-project/modulectl/internal/common/errors"
+	"github.com/kyma-project/modulectl/internal/common/types"
 	"github.com/kyma-project/modulectl/tools/io"
 )
