+++ internal/service/filegenerator/filegenerator_test.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
-	commonerrors "github.com/kyma-project/modulectl/internal/scaffold/common/errors"
-	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
+	commonerrors "github.com/kyma-project/modulectl/internal/common/errors"
+	"github.com/kyma-project/modulectl/internal/common/types"
 	"github.com/kyma-project/modulectl/internal/service/filegenerator"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
