+++ internal/service/filegenerator/reusefilegenerator/reuse_filegenerator.go
@@ -3,8 +3,8 @@ package reusefilegenerator
 import (
 	"fmt"
 
-	"github.com/kyma-project/modulectl/internal/scaffold/common/errors"
-	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
+	"github.com/kyma-project/modulectl/internal/common/errors"
+	"github.com/kyma-project/modulectl/internal/common/types"
 	"github.com/kyma-project/modulectl/tools/io"
 )
