+++ internal/service/moduleconfig/moduleconfig.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"path"
 
-	"github.com/kyma-project/modulectl/internal/scaffold/common/errors"
-	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
+	"github.com/kyma-project/modulectl/internal/common/errors"
+	"github.com/kyma-project/modulectl/internal/common/types"
 	"github.com/kyma-project/modulectl/tools/io"
 )
